test/requirements.txt
datasets==3.2.0
requests==2.32.3
tqdm==4.66.5
fastapi==0.115.5
uvicorn==0.32.0
google-generativeai==0.8.3
